Fix the bug where if you're viewing the 'unread' filter, then as cards get auto-read they disappear.
Introduce notion of pendingFilters and filters. The latter is what is actually used for calculating what cards to show. The former is what is updated as filter info changes. They are only synchronized when explicitly told to. That synchronization happens every time refreshCardSelector is called before dataIsFullyLoaded (because that's likely information actually important to show), but after dataIsFullyLoaded synchronization only changes when the collection parameters (active set names and fitlers) change, which optimizes for consistency in collection, and fixes the odd behavior. Fixes #109.