Fix the bug where `c/starred/_` canonicalized to `c/_` when there were only fallback cards.
The logic said that those cards were orphans so should have no filters, but in that context they aren't.

Also, changed the behavior in last commit, because now that the fallback card is not considered an orphan the old logic would have worked just fine for `c/reading-list/_` with no items, too.

Part of #197.