MANDATORY MAINTENANCE TASK ADDED. Creat a todo field on cards. Part of #200.
Add a todo field on cards that operates similarly to the existing notes field, but differs in that it is not displayed in the card info panel, and when set communicates that the card has some outstanding TODO to be done.

Because the existing notes field on existing cards has been used as a de facto todo field, the maintence task copies over existing notes fields to todo field.

This commit simply adds a default todo field in new cards and a maintencne task; it doesn't actualy allow setting or doing anything with it.