Fix a bug introduced in 2d65202, so that URLS without a trailing "_" work on load. Fixes #204.
The problem was that 2d65202 changed it so that a requestedCard of "" meant "default card in that collection, equivalent to '_'". The problem was that logic in showCard compares the current activeCardId to the requestedCard, and if they're the same it doesn't do the logic to convert it from a placeholder to a real card. The intial state has '' for the activeCardId, so on the very first load it was bailing.

The solution is to have under the covers the requested card be "_" even when the trailing part of the URL is "", so that it's clear it's not just initial value. Then we just take care to never show it to users in the URL.