Instead of picking the first tweet in sorted order, pick randomly from the entire set that has the equivalent sort value as the first item.
This introduces a bit more "spice" and unpredictability into the tweet order.

In general, the equivalence classes will have a single item once in steady state, but at the beginning there are lots of cards that all haven't been touched since the import and also ahven't been tweeted.

Part of #212.