Remove zombie span notes that are introduced when you backspace at a paragraph boundary.
Part of #228.

Note, however, that there's still an oddity: the HTML that shows in the content textarea is right, but the actual HTML in contenteditable itself is wrong, so if you hit backspace and then space, then it will add an nbsp; because it's at the end of the textnode and the actual span is still there.

Probably needs something to actually force content editable to actually take the transformed content or something.