Support union filters, which are filters that combine multiple concrete or inverse filter names like so "starred+in-reading-list+unpublished".
By making them literally just be strings with that delimiter, the rest of the machinery doesn't have to change much; they can still be chained with normal filters and unioned with them.

Have to wire through allCardsFilter anywhere that calls into combinedFilterForFilterDefinition so that we can support union sets that include inverse filter names, because they must be expanded.

Part of #139.