Reintroduce partial string matching by indexing all substrings of words, too.
This leads to a 10x memory increase in the index.

Performance returns to being bad, but it turns out it's not because of the matching (that's trivial). It's because without partial matching, the expensive template instantiation had to be done much more rarely, and that's what the performance problem was before.

Part of #93.