Change the way query debouncing works, to try to wait until after the user has stopped typing, instead of some period of time after they STARTED typing.
Without this change, if you were writing a long enough word you'd run into the query being updated, which would freeze it.

Set the debounc to an interval short enough that once you stop tying the results update quickly, but long enough that in the course of normal typing you don't run into it often.

Part of #93.