Fix a heisen-failing test in the last commit.
Make it so we use the new map.get('key', defaultValue) getter so userIsAdmin doesn't fail if the user has an object but DOESN'T have an admin key.

Part of #270. Part of #190. May have broken when certain accounts had an object created to work with 7501b3b3deab15fbff58478727e95890fc17cc46.