Switch to a totally different calculation for semanticDifference.
Instead of calculating the amount of difference between tfidf of their words, instead just return 1.0 minus the number of terms that are in both fingerpints, over the number of total terms in the fingerprints.

This makes sense because the key terms, even with no cap on fingerprint size, is only 11% across the corpus. So that meant that the prior calculation basically reduced to the size of the tfidfs. That penalized cards that had  a large number of rare words, and rewarded cards that were very average. That meant that most of the related cards were just the same average cards.

This approach instead focuses on the fact that the tfidf doesn't matter for the term, other than selecting the most distinctive terms for the card. What matters is how much the most distinctive terms overlap.

Part of #280.