Tie in a cloud function so that when a card links are updated, page rank is recomputed for all cards in a cloud function and stored in card.card_rank.
There is no maintence task, because it will run automatically on the first card edit, so the amount of time an old database would be out of date is very limited.

Not committed on main branch because storing the updates to all of the cards that change takes TONS of time. Basically every card's page rank changes every time a link is changed, which is too slow. Part of #151.