Fixed a bug where screenshots always showed red links for unpublished cards.
The local logic assumes that unpublished cards won't be downloaded--if they are, then the user has access to see them.

The screenshot service injects the linked cards into the page directly. It didn't used to verify that the cards were published. So the card was injected, the renderer saw they existed and were unpublished, and they showed up as red.

Now, we check the cards are published when fetching them to inject.

Increment the screenshot version because this will change all of the rendering for any cards with unpublished links.

Part of #314.