Small tweak, necessary to work in a (broken?) implementation of jsdom in testing.
When hoisting a text run into a wrapped element, we set the new element's innerText. But jsdom ignored that inner text, discarding it when hoisted.

Here we just use innerHTML as the setter. This avoids a bug in jsdom so we can exercise the actual in-browser behavior better in next commit. Part of #234.