Make it so that admins can manually add a uid from card-editor collaborators or editors.
It has to be an admin, because a) they need the uid in the first place, which implies access to the firebase console. And b), otherwise it would be anyone who can edit a card, but that's not easy to test generically without knowing the card they might be able to edit. And if we didn't protect who could create authors entries, then anyone could spoof a given uid to publicly divulge their displayName and photoURL next time they loaded the app.

Part of #331.