Update a few tests.
Turns out they were failing (correctly), but not because the LOGIC was correct, but because they were hitting the 1000 rules limit.

These tests SHOULD have broken wehn bob was added as author to the cards, which allows him to edit any card any way he wants.

Switch the tests to use just a generic user.

Part of #286. Noticed while working on #331.