Add a getPermissions() function that allows permissions checks for multiple keys to be more efficient.
It calculates the set of keys for the user a single time, saving many ops for cases where you could have either of two (or three) different permissions, e.g. 'edit' and 'editSection'.

This is only possible as implemented because we verify that all permissions keys are true (for user objects we don't formally do that, but the UI doesn't give any affordances to add one).

Part of #286. Part of #331.