MAINTENANCE TASK REQUIRED. Switch to saving a card's editors within a permissions.editCard object on the card.
This is a teensy bit odd, but it will allow adding viewUnpublished to a single card and user to work, and will also be a more similar system to what we'll use for #276.

Note: the maintenance task is the same one as the last one added, but with its behavior changed. This is OK because the earlier maintenance task hasn't been run in production yet.

Part of #331.