Made the drag/drop way less finicky.
The problem is the drop target was only below the line in the middle. Changed it so that the drop target is larger and also centered on the space in the middle.

This makes it WAY WAY less bouncy.

The one downside is that the line in the middle of the cards doesn't show (it would be overlapping the top card). But that's not that big of a deal because the space between cards makes it unambiguous where the drop will go.

Part of #335.