Fix the bug introduced in the last few commits where cards in `c/everything/orphaned/` that were content cards would have their orphan URL.
In the process, simplified some of the odd canonicalization logic and made it simpler. It SEEMS to work in all of the cases I could think to ad hoc test.

Direct links to orphan cards will still only show that card. But links to oprhanded cards in the context of a collection that is non empty will keep the collection and show the other cards. This makes sense--you don't want to see all of the other orphaned cards if you are viewing one, because the other orphaned cards shouldn't be revealed unless someone goes looking for them. Part of #358.