Make it so that changes to card opbjects will never have 'null' for timestamps.
docChanges gets THREE updates when cards are changed, and at the beginning serverTimestamps are null.

This hasn't been a problem before, but if we want to open the cards for editing immediately and peek at those timestamps, they'll be null.

Tell firebase to estimate them. https://medium.com/firebase-developers/the-secrets-of-firestore-fieldvalue-servertimestamp-revealed-29dd7a38a82b

Part of #353.