MASSIVE commit, with TWO required maintenance tasks. Very likely broke something, although I've verified most of the behaviors.
Switch the underlying data model from `links`, `links_inbound`, `links_inbound_text`, `links_text` to instead be primarily `references` where it's a map of cardID to link-type to their text. The system also maintains a `refernces_sentinel` (and its inbound inverse), which is just a map of cardID to true, and is there entirely so that firestore can be queried for cards that are linked to by other cards.

You have to run 'links-to-references' maintence task to generate references, and then run 'update-inbound-links' to set the inbound references. Both of those require the app to be in maintenace mode.

Part of #354.

Introduces the slightly odd division of labor between modifyCard and editingCommit that #368 is tasked with fixing.