MAINTENANCE TASK REQUIRED. Rename 'references' --> 'references_info' and 'references_sentinel' --> 'references'.
This changes the behavior of the maintenance task introduced in 5d90b9638de71c79d167eb2ad20c44071751c3ba. If you already ran that, you need to reset from a backup and then run this maintenance task again. That task was never run in production, so I didn't bother doing a transfer from the old way.

Renaming these fields makes it more clear the expected behavior of each. Detailed comments on their behavior to come in the next couple of commits. Part of #354.