NEW MAINTENANCE TASK: add-font-size-boost.
Cards get a font_size_boost object. If there's a key set, that denotes that the given field on the card should be boosted so the font-size is 1.0 + font_size_boost, with missing keys denoting 0.0.

All cards have a blank boost map, and it's not yet used to render. Part of #371.