Card semantic fingerprints include bigrams as well.
The bigrams are weighted less strongly than underlying words so they don't compete with them that closely. Because there are now more items, we've increased the size of fingerprints somewhat.

Gut checking lots of different fingerprints, it looks like the bigrams that are sorting to the top are very nice and should lead to better matches.

This likely made the performance somewhat worse: larger fingerprints and also more words to match against.

Part of #365.