Add a 'limit/10' filter.
This is a very special filter, because it can only operate once the final sorted collection is known.

the actual configurable filter operates as a noop configurable filter in the main pipeline.

CollectionDescription learns how to parse out any limit filters itself. And Collection checks the description for any limits. If there is one, then it fudges the behavior of filteredCards, numCards to figure out the limit beforehand, which requires a sort to happen first.

Fixes #387.