Introduce the 'similar' filter, which operates on cards that are semantically similar to the given key card.
It's an odd 'filter' in that it doesn't actually do any filtering, but is rather used for its indirect sorting.

It would notionally be better as a sort, but sorts don't have machinery for configurable sorts so :shrug:

The cardID can have a `+` prepended which communicates that the key card should be included.

Often used with limit, e.g. `c/similar/c-123-45678/limit/5/`

Part of #310.