MAINTENANCE TASK REQUIRED: flip-auto-todo-overrides
Flip the meaning of true/false in card.auto_todo_overrides, so true means "the TODO is marked as done no matter what the auto todo function says" and false means "the TODO is marked as not done no matter what the auto-TODO function says."

The previous logic was reversed from the way that the FILTER_CONFIG was oriented (where a true from that function is "good"/"done"), which was needlessly confusing.

Part of #384.