Make FingerprintGenerator filter out cards that are not body cards by itself, instead of doing it upstream.
This difference was one of the reasons that the similar card filter would give different results--it was passing all cards, but the fingerprint generator for everything else was using only body cards. By having FingerprintGenerator filter the cards itself, it's more resilient to giving similar answers no matter what.

Part of #400.