Fix a serious performance regression while editing cards. Fixes #411.
The problem was that selectCollectionConstructorArguments included editingCard, which doesn't change often, except when a card is being edited, in which case it changes with every keystroke.

That meant that selectActiveCollection was being invalidated with every keystroke, which meant that things like card-drawer itemsToGhost were being invalidated with every keystroke, which could be expensive.

Most places don't require editingCard to be passed, so create a selectCollectionConstructorArgumentsWithEditingCard that can be used in cases where you really do want the editing card, like for info panel (where it's a feature, not a bug that the similar cards update within a second or so after typing stops).