Introduce a layer of indirection around selectEditingCard, by also introducing selectEditingNormalizedCard.
The former returns the card, WIHTOUT normalized text properties. The latter includes NLP normalized text properties, but only updated on a delay. Everything that needs NLP properties uses selectEditingNormalizedCard transitively now.

Currently selectEditingNormalizedCard is just an alias for selectEditingCard, but with this setup, it's possible to now start having editingCard not be literally expanded with extracted text content in EXTRACT_TEXT_LINKS. Part of #412.