selectEditingNormalizedCard grows to have custom memoization that only returns a new version when cardExtractionVersion changes.
That number changes when EXTRACT_CONTENT_LINKS is fired (or editing is started or stopped). This means that evertyhing that's downstream of nlp properties on editing card will now not need to worry about rate limiting on card access.

The machinery that used to literlaly change the nlp properties of editing card at EXTRACT_CONTENT_LINKS time now just extracts the links, and then signals to the selectEditingNormalizedCard that it should reclaculate. Part of #412.