selectCards now does a more complex selection including the fallbackMap for each card.
We have to use a customEquality function so that most cards aren't reprocessed, that doesn't check if the zipped array is the same each time (it will change every time) but if the underlying card and its fallbackMap (very likely null) do.

Verified by flipping the debug code on that it doesn't lead to more normalized properties being calculated than before.

cardWithNormalizedTextProperties is sent the fallback map, but doesn't do anything with it yet.

Part of #412.