Add strong text as a type of normalized text field in nlp.
This extracts all of the text that is in strong tags in body as separate runs. It essentially doubles the emphasis of strong text (since it was already extracted directly in body), but also because the strong text runs are likely short, it makes them more likely to be extracted as large-ngrams.

Part of #399.