Fix a bug introduced in 1a0972973c7531921f4f5bc4613ed7e9d8705eb4 and 1eb05cb9c929fbc450eb111c109d7db64d4fee06.
ngrams is supposed to return ngrams of literally only X size, but it was returning smaller ones for small text. And, it was returning special full strings too often.

Changed it so the special long ngram is done in wordCountsForSemantics and WHOLE_NGRAMs are done there too, so they don't double count. Part of #399.