Fixed the bug where new cards would have their title blurred when opened for editing.
The problem is that once updateCards is called with the new card, selectCollectionConstructorArguments is now different (cards is different), which would lead to a change being made on card-renderer, which led to the selection being lost.

The fix is to allow noop changes in the common case where there are no primary reference blocks to be detectable as not changing by using a singleton.

This was originally broken in 5c48184e45d982b46d98838a02848f248d97e0e0

Part of #399.