Make it so when navigating to a new card, we don't try to open it for editing if it's a card with autoSlug.
This is a kind of hack for a problem where if you make a card, like a concept card, that will be autoslugged, and the card immediately opens for editing before the autoSlug has bene applied, then if you save the card you'll overwrite it and ignore the new slug.

... This is kind of a hack.

Part of #399.