Generalize the suggestedConcept machinery to consider more types of existing references as invalidating a possible concept suggestion.
Before, if you added a synonym reference, then it almost certainly brought in text that would lead to that concept ranking high, which would lead to a suggested concept of the card you already referenced. This new machinery allows types that are conceptually sub-types of the concept reference to be considered an existing concept reference.

Part of #401.