xMake it so there's a minVersion for every maintenance task.
When a maintenance task is run, it stores the timestamp and also the current MAINTENANCE_TASK_VERSION.

This will allow us to detect, based on the version stamp of INTIAL_SET_UP task, which maintenance mode tasks don't need to be run, because this instance of the app was set-up after that maintenace task. Part of #217.