Mess with the logic of selection saving/restoring.
Instead of literally saving and restoring a selection, it remembers the offsets in the node for visible text.

This is more finicky (it needs the container element to literally be the same, for example), but it THEORETICALLY would allow selections to be restored even when the underlying HTML has changed (so long as no user-visible characters are added).

Part of #449. Part of #364.