Revert "Mess with the logic of selection saving/restoring."
This reverts commit 06e2b876097a16b15c7b33fe0956daae9338c7fd.

It caused a significant performance regression after editing a few cards.

Fixes #452. Part of #449, part of #364. The code in this commit is useful, but will have to be redone because the rendering code in the card-renderer that reuses the same element will need to be really careful to not add too many event handlers.