Refactor how 'subtle' works on tag-list.
When subtle is set, it sets a subtle property on each tag-chip.

When subtle is set on a tag-chip, it forces on a subtle filter style, overriding any tagInfos style.

When you hover over a filtered tag-chip it animates to no filter.

You can provide subtle manually on a tag-chip, or via the tagInfos for that tag.

Part of #174.