Refactor card.nlp to maintain information on each run in one place.
It used to be card.nlp had three properties: normalized, stemmed, withoutStopWords, each of which was a map of field name to an array of strings.

Now, it's nlp as a map to field name, each of which is an array of runs. Each run has original/normalized/stemmed/withoutStopWords values.

This will make it easier to do processing that requires multiple versions of the same string.

Performance doesn't appear to have degraded. Everything APPEARS to work, but I'm sure something must have broken.

Part of #436.