Trim down the list of STOP_WORDS by about half. Remove all of the words from "once" and downward, as well as "I", "Me", "My", "Mine", "Myself", and 2nd and 3rd person variants.
This is a kind of hack, but it should be pretty effective at fixing the most annoying issue with stop words, where `know-how` matches ANY instance of `know` erroneously.

This list is way better anyway, the other one was way too expansive and missed important things. The "I"/"Myself" stuff in particular was challenging, it made it impossible to search fo re.g. "insights about myself" (a string that is in lots of working notes) and have good results.

Part of #436.