Refactor how configurable filter funcs are passed arguments to be easier to extend.
Every configurable filter func needs to be passed information that ANY of them needs. Before that meant adding a new argument that all of them got, and wiring that through the whole pipeline, which is finicky and error prone.

Now, filter funcs are given card and extras, an object with named properties for each. This will make it way easier to wire through new information with just a few call sites differing.

Performance doesn't seem to regress.

It's possible that some of the configurable filter funcs broke in this change.

Part of #495.