Simplify the auto-merging logic. When the undelrying card updates, update it!
This gives much closer to the desired logic but much more simply. The reason you have a snapshot is simply because you need to notice when it changes so you can diff the user's edits from the old snapshot to apply to the new snapshot.

This has behaviors like "if the user hasn't yet updated title themselves, update it... but not if they've updated it."

Two problems: 1) the naming of auto-merging is wrong and weird now. 2) There's no way for a user to see what edits happened underneath that they might want to discard their own for.

Part of #332.