Change the implementation of ngramWithinOther to not use memoized regular expressions, preventing OOMs that were happening production every time a card was edited.
ngramWithinOther used to use (memoized) regular expressions to check for a word-boundary match in the ngram. This lead to a RE being cached for every single ngram in every single card. Even though it was only 60,000 ngrams or so it was still causing non-trivial memory usage.

This change does affect performance a bit, going from 5.3s to load to 7.4s. But now you can actually edit cards in production again so it's worth it.

This doesn't fundamentally fix the OOM issue--we're still dancing on the edge of the OOM cliff with this architecture. It just staves off the inevitable a little longer.

Part of #522.