Revamp suggestedConceptReferencesForCard to not rely on the fingerprint.
Before, it implicitly relied on the fact that possible concepts would always be boosted to the point they showed up in the 50 fingerprint items.

But that meant hat concepts, and proto concepts, had to be hella boosted to make sure they showed up in suggestions--and then once they're explciitly added they would DOMINATE fingerprints.

This pipeline makes it so concepts will be found and suggested even if they don't make it into the fingerprint, allowing much more tweaking of fingerprint generation without breaking the concept suggestion pipeline.

Part of #533.