Make it so drag to reorder works again (REORDERING_DISABLED is still true though so it still doesn't).
drop targets for card thumbnail spacers have the ID of the card they are affiliated with, instead of the collection-specific index (which was finicky!)

This radically simplifies reorderCard, because it is just a thin wrapper around modifyCardWithBatch.

This simplified logic will make it much easier to support more complex reordering, e.g. in reversed things. Part of #199.