Fix sortOrderIsDangerous so that it no longer reports numbers that go negative as being dangerous.
It was using Number.MIN_VALUE erroneously--that's the smallest value that's not zero. But we wer elooking for the most negative number, which is Number.MAX_VALUE * -1.

The maintenance task to assign sort_orders very specifically puts numbers right up to 0.0, so any additiional card would go negative and be illegal, but actually have huge amounts of range left to it.

Fixes #542.