Fixed it so when opening find dialog, the similar cards shown are similar to the editing card, not its last saved state.
This fixed a long-standing bug that only recently became more noticeable when the default set order changed. The find dialog, when first opened, should use the editingCard's state, but wasn't, to show the default similarCards list. New cards with blank content aren't similar to anything, so you just get the default set. It used to be the default set was sorted in a way that recently touched cards were at the top, which was often enough kind of what the user wanted, so they didn't notice. But now, we correctly use the editing card as well for filters running in the find dialog, so we properly show cards that are similar tot he editing card's current content, not just waht it was when it was last saved.

Fixes #545.