Make it so when the find card dialog is open generically it just shows recent cards.
Before, if you opened the find card dialog on the first card (e.g. a section-head) you'd just get a random set of the boring about cards, because nothing is similar to section-head.

Now we only do that when the find dialog is being used for finding a card to link.

Fixes #555.