Make it so that the diff / unsaved changes code uses selectEditingNormalizedCard.
This card changes far less often (it delays after typing ends). Before, on every keystroke we were normalizing the card HTML TWICE just to detect if there were any changes to warn about whether there were unsaved changes in the UI. This could create some lag on keystrokes. The trade off is that it MIGHT not have a race where if you try to save or cancel a card editing before the normalized properties have been calculated it might not warn you properly.

It's a small improvement but not huge. Part of #564.