Make it so generateCardDiff by default doesn't normalize HTML.
This removes a very expensive (and memory-heavy) operation on the hot path when editing text fields in a card.

It's a safe assumption that we can skip the diff for everything except the final diff used before saving, because right now (before #503 is fixed) if body is changed, the entire body is changed to the diffed value, there is no intra-field-diffs.

This makes typing much faster and is a better solution than #567.

Part of #564.