Make it so suggestedConceptReferencesForCard doesn't take conceptCards as an argument, leading to 7x speed up on key tasks.
Part of #580. Part of #571.

suggestedConceptReferencesForCard used to accept condeptCards and concepts, but actually conceptCards was only used for its ID anyway, which was already in concepts. By changing its signature to only take the card and the concept map, there are SIGNIFICANT speed ups. This is likely for some mix of two reasons: 1) conceptCards, despite extra layers of caching introdced in recent commits, was still changing every so often; the concepts map is changing less often so it can use cached results more often. 2) getConceptCardForConcept is a linear scan through all concept cards, which was run in a hot loop for every concept match and is now omitted for a simple map lookup.

In either case, this is simpler and signficantly faster.