modifyCards now calls updateCardsWithOverlay on success.
We've manually verified the machinery does succesfully note the cards are equal and drop later no-op updates.

Note that currenty the firestore snapshot runs BEFORE the overlay is run (since it's cached and runs syncronously before the server ACKs), and then immediately after with the updatedCardsOverlay.

This means the call with updatedCardsOverlay is currently mostly a no-op, but that won't be true in a few commits when we do a local version of updateInboundLinks in the overlay creation.

Part of #581.