Make it so forking also updates inboundLinks manually.
In the (edge) case where some of the outbound references are to cards that this user cannot see, then we either offer to abort or to remove those references in the fork.

Note that there are edge cases where some people with partial permissions can't fork a card, but that appears to predate this refactor.

Part of #585.