Switch from `static get styles() {` to `static styles = [`
This is the preferred style for Lit and is now possible given that eslint is upgraded in 8c8c3f33b69dd8262390cf3c8db66d91dfb81563.

This is either a teeny bit faster (if Lit only calls this a single time) or a lot faster (if it's called on every instance, and the comparions is for deep equality, since the former style would generate a NEW array each time). But in any case it's the preferred style.

This is mostly a very mechanical, small update and re-indent.

Part of #453.