Create a system of serverTimestampSentinel that allows literal Timestamps to be used as serverTimestamp sentinels.
serverTimestamp returns a FieldValue, not a Timestamp. defaultCardObject returns a CardUpdate, not a Card, because of that. But some of the values created by defaultCardObject need to be able to be treated like a Card, e.g. passed to finishers.

This system allows us to vend a normal, currentTime timestamp... but also keep track of the fact that they're notionally a serverTimestamp, to be converted later.

Later, you pass the card update to installServerTimestamps(). But if you forget, it's not a big deal, because instead of getting a server-set currentTimestamp, you get a client-set one. :shrug:

Not used anywhere yet.

Part of #601.