Remove reselect-tools, resolving a noImplicitAny
I haven't used it in awhile, the library hasn't been touched for years, doesn't have types, and it required a weird import (reselect-tools/src) that made it hard to figure out how to do an override set of types.

I might regret this in the future, but it seems cleaner to just get rid of it.

Part of #601.