Make it so auto-todo-filters apply even for cards that don't get the auto-todo.
Before, working-notes would all match published, even if they werent' published... because being published isn't a requriement. But that's super weird! Because factually they are unpublished.

This logic change almost certainly broke something... it must have. But the does-not-need-published should work correctly for that use case. And cardTODOConfigKeys is really the main meat of which auto TODOs are required.

Part of #620.