Make the working-notes default tab also add unpublished filter.
This is a no-op in the common case. But in production we'll soon make it so there are published working-notes cards... which should show up in a separate tab, not this one. So this makes sure only unpublished ones show up in this tab... that is the section only shows up for non-editors.

Part of #620.