Fix Point.card_created.
Before we were using the nanoseconds, which is NOT the nanoseconds since the Unix epoch, but instea is nanonseconds on top of the seconds.

Now we actually store the milliseconds as the comment says.

This is not worth reindexing for because this field isn't used for much anyway.

Part of #646.