change reference blocks so see-also shows up for inbound-see-also too.
This requires changing the "cards that link here" to skip see-also inbound links, since it's already
showed above.

This makes see-also be a reciprocal type which is conceptually correct... but also means that it can be a
bit confusing, if you want to remove the connection it's hard to tell if you're looking at the primary or
reciprocal card... so you might hit edit and then be surprised to not see a see-also reference to remove.

Part of #670.