Allow the CardSimilarity endpoint to take a last_updated.
If provided, then error of code `stale-embedding` will be returned if the embedding is older than the
card.updated.

This will allow the client to determine that the embedding isn't  yet updated, and try again. This will
happen for example right after a card is saved, before it is re-embeddeded and stored.

Part of #670. Part of #646.