Plug in suggestionsForCard in a silly way, so it's called when you view a new card.
This way, it's a simple VERBOSE = true change in one place to get the suggestions pipeline going.

This is just a way to play around with the pipeline, this is likely not the right place to trigger this.

For most users this will basically do nothing, because it will see they don't have edit access and return immediately.

Part of #670.