When a card is being edited, use tfidf similarity.
There's a current bug where you create a card, and it gets a '' embedding (which returns very generic results). While editing, you see those generic results, even as you update.

In the future we'll be live fetching the actively edited card, but for now just fall back on tfidf pipeline.

Part of #646.