Set the LAST_UPDATED_EPSILON much tigher.
Before, it was 10 seconds, allowing a lot of slop between servers. But that broke a common case: create-a-card, and then within 10 seconds paste and save.

The client would then ask for the new similarCards, and it would still be based on the old empty content embedding, and the server would happily report it as current.

That would lead to bizarre similarity for that card (until another card was created/updated, when that cache would be blown away in the client and refetched and would actually be accurate).

Part of #646.