Add a reverse argument to makeReferenceSuggestion
Before, we flipped the key/supportingCard when it should be reversed, but that would lead to a weird
situation where the key card might not be the current card, which would be confusing when swapping
between two cards on either side of the suggestion, where it might not shift.

Now we have the variant where keyCard is always the current card.

Part of #670.