Calculate suggestors in parralel.
This allows suggestions from the fast suggestors to stream in very quickly, making it feel way more
responsive.

The downside is log message are interleaved in a confusing way, and the better techniques (which tend to
be more intensive) tend to show up at the end of the list.

Part of #670,