For autoSlug, break it into `primary` and `prefixed`.
The former behavior is `primary`: it tries to slug the card with its title, and falls back on using a prefix.

`prefixed` only tries with the prefix.

We then use htat for `work` cards. The reason is that a number of work cards have titles that are generic enough (e.g. Situated Software) that I might want to use them for a concept card later.

Part of #465.