Move where cardFinsihers and fontBoosts are done to be within modifyCardWithBatch.
Before, the editor was responsible for running it. But that meant it was possible to have cards who wouldn't have their finishers run.

For example, a suggestor to convert to quote card wouldn't run the finisher, so the title would remain with the old title.

But now, everything right before getting modified gets the final finsiher.

Part of #465.