Fix it so when you create a new card it navigates to it in the current collection.
ef476f53966e298bbf8034720522ff7692b0153f was a partial fix to ba3f7c8, but it was partially broken because it modified the supposedly immutable loadingCardFetchTypes of the previous state... oops! This mean that the cardsLoaded selector saw object equality and assumed no change, when in reality it had changed, so dataIsFullyLoaded erroneously still returned true.

Part of #681.