Make the bulk-fetch for reindexCardEmbeddings also ask for the card_id.
Hopefully this will fix #691. I believe currently the cardsInfo is a map of {'':<last-embedding>} because we don't tell qdrant to include the card_ID in payload.

It's kind of hard to debug this with long cycle times, and if that's true, this optimization NEVER worked. (Or maybe Qdrant changed its behavior and previously returned all payload fields and now listens to the config)?