Switch to not try/catch around updatePermissions.
When that happens today, as happens reasonably often in #690, this swallows the error and makes it harder to debug.

There shouldn't be an error under normal flows I don't think, and if the permissions doc doesn't exist the snapshot will have an exists() false, and return undefined from data(), which is already handled.