Attempt to fix a login bug with merging accounts.
Before this, if you logged in to a previously existing account, the confirmation prompt would be auto-closed by Chrome because the popup was still showing, and print: `A window.confirm() dialog generated by this page was suppressed because this page is not the active tab of the front window. Please make sure your dialogs are triggered by user interactions to avoid this situation. https://www.chromestatus.com/feature/5140698722467840`

Because it would be auto-canceled, the sign in would be stopped according to our code.

This change attempts to change timing so the popup is closed when that happens. Not sure if it works.

Part of #693.